+++ skrl/models/torch/gaussian.py
@@ -35,21 +35,21 @@ class GaussianMixin:
 
         :raises ValueError: If the reduction method is not valid.
         """
-        self._clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
+        self._g_clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
 
-        if self._clip_actions:
-            self._clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
-            self._clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
+        if self._g_clip_actions:
+            self._g_clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
+            self._g_clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
 
-        self._clip_log_std = clip_log_std
-        self._log_std_min = min_log_std
-        self._log_std_max = max_log_std
+        self._g_clip_log_std = clip_log_std
+        self._g_log_std_min = min_log_std
+        self._g_log_std_max = max_log_std
 
-        self._distribution = None
+        self._g_distribution = None
 
         if reduction not in ["mean", "sum", "prod", "none"]:
             raise ValueError("Reduction must be one of 'mean', 'sum', 'prod' or 'none'")
-        self._reduction = (
+        self._g_reduction = (
             torch.mean
             if reduction == "mean"
             else torch.sum if reduction == "sum" else torch.prod if reduction == "prod" else None
@@ -79,24 +79,24 @@ class GaussianMixin:
         log_std = outputs["log_std"]
 
         # clamp log standard deviations
-        if self._clip_log_std:
-            log_std = torch.clamp(log_std, self._log_std_min, self._log_std_max)
+        if self._g_clip_log_std:
+            log_std = torch.clamp(log_std, self._g_log_std_min, self._g_log_std_max)
             outputs["log_std"] = log_std
 
         # distribution
-        self._distribution = Normal(mean_actions, log_std.exp())
+        self._g_distribution = Normal(mean_actions, log_std.exp())
 
         # sample using the reparameterization trick
-        actions = self._distribution.rsample()
+        actions = self._g_distribution.rsample()
 
         # clip actions
-        if self._clip_actions:
-            actions = torch.clamp(actions, min=self._clip_actions_min, max=self._clip_actions_max)
+        if self._g_clip_actions:
+            actions = torch.clamp(actions, min=self._g_clip_actions_min, max=self._g_clip_actions_max)
 
         # log of the probability density function
-        log_prob = self._distribution.log_prob(inputs.get("taken_actions", actions))
-        if self._reduction is not None:
-            log_prob = self._reduction(log_prob, dim=-1)
+        log_prob = self._g_distribution.log_prob(inputs.get("taken_actions", actions))
+        if self._g_reduction is not None:
+            log_prob = self._g_reduction(log_prob, dim=-1)
         if log_prob.dim() != actions.dim():
             log_prob = log_prob.unsqueeze(-1)
 
@@ -111,9 +111,9 @@ class GaussianMixin:
 
         :return: Entropy of the model.
         """
-        if self._distribution is None:
+        if self._g_distribution is None:
             return torch.tensor(0.0, device=self.device)
-        return self._distribution.entropy().to(self.device)
+        return self._g_distribution.entropy().to(self.device)
 
     def distribution(self, *, role: str = "") -> torch.distributions.Normal:
         """Get the current distribution of the model.
@@ -122,4 +122,4 @@ class GaussianMixin:
 
         :return: Distribution of the model.
         """
-        return self._distribution
+        return self._g_distribution