+++ skrl/models/torch/multivariate_gaussian.py
@@ -28,17 +28,17 @@ class MultivariateGaussianMixin:
         :param max_log_std: Maximum value of the log standard deviation if ``clip_log_std`` is True.
         :param role: Role played by the model.
         """
-        self._clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
+        self._mg_clip_actions = clip_actions and isinstance(self.action_space, gymnasium.Space)
 
-        if self._clip_actions:
-            self._clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
-            self._clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
+        if self._mg_clip_actions:
+            self._mg_clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
+            self._mg_clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
 
-        self._clip_log_std = clip_log_std
-        self._log_std_min = min_log_std
-        self._log_std_max = max_log_std
+        self._mg_clip_log_std = clip_log_std
+        self._mg_log_std_min = min_log_std
+        self._mg_log_std_max = max_log_std
 
-        self._distribution = None
+        self._mg_distribution = None
 
     def act(
         self, inputs: Mapping[str, Union[torch.Tensor, Any]], *, role: str = ""
@@ -64,23 +64,23 @@ class MultivariateGaussianMixin:
         log_std = outputs["log_std"]
 
         # clamp log standard deviations
-        if self._clip_log_std:
-            log_std = torch.clamp(log_std, self._log_std_min, self._log_std_max)
+        if self._mg_clip_log_std:
+            log_std = torch.clamp(log_std, self._mg_log_std_min, self._mg_log_std_max)
             outputs["log_std"] = log_std
 
         # distribution
         covariance = torch.diag(log_std.exp() * log_std.exp())
-        self._distribution = MultivariateNormal(mean_actions, scale_tril=covariance)
+        self._mg_distribution = MultivariateNormal(mean_actions, scale_tril=covariance)
 
         # sample using the reparameterization trick
-        actions = self._distribution.rsample()
+        actions = self._mg_distribution.rsample()
 
         # clip actions
-        if self._clip_actions:
-            actions = torch.clamp(actions, min=self._clip_actions_min, max=self._clip_actions_max)
+        if self._mg_clip_actions:
+            actions = torch.clamp(actions, min=self._mg_clip_actions_min, max=self._mg_clip_actions_max)
 
         # log of the probability density function
-        log_prob = self._distribution.log_prob(inputs.get("taken_actions", actions))
+        log_prob = self._mg_distribution.log_prob(inputs.get("taken_actions", actions))
         if log_prob.dim() != actions.dim():
             log_prob = log_prob.unsqueeze(-1)
 
@@ -95,9 +95,9 @@ class MultivariateGaussianMixin:
 
         :return: Entropy of the model.
         """
-        if self._distribution is None:
+        if self._mg_distribution is None:
             return torch.tensor(0.0, device=self.device)
-        return self._distribution.entropy().to(self.device)
+        return self._mg_distribution.entropy().to(self.device)
 
     def distribution(self, *, role: str = "") -> torch.distributions.MultivariateNormal:
         """Get the current distribution of the model.
@@ -106,4 +106,4 @@ class MultivariateGaussianMixin:
 
         :return: Distribution of the model.
         """
-        return self._distribution
+        return self._mg_distribution