+++ skrl/utils/model_instantiators/jax/gaussian.py
@@ -8,8 +8,9 @@ import jax
 import jax.numpy as jnp  # noqa
 
 from skrl.models.jax import GaussianMixin  # noqa
-from skrl.models.jax import Model
-from skrl.utils.model_instantiators.jax.common import generate_containers
+from skrl.models.jax import Model  # noqa
+from skrl.utils.model_instantiators.jax.common import one_hot_encoding  # noqa
+from skrl.utils.model_instantiators.jax.common import convert_deprecated_parameters, generate_containers
 from skrl.utils.spaces.jax import unflatten_tensorized_space  # noqa