+++ skrl/utils/model_instantiators/torch/categorical.py
@@ -8,7 +8,8 @@ import torch.nn as nn  # noqa
 
 from skrl.models.torch import CategoricalMixin  # noqa
 from skrl.models.torch import Model
-from skrl.utils.model_instantiators.torch.common import generate_containers
+from skrl.utils.model_instantiators.torch.common import one_hot_encoding  # noqa
+from skrl.utils.model_instantiators.torch.common import convert_deprecated_parameters, generate_containers
 from skrl.utils.spaces.torch import unflatten_tensorized_space  # noqa