+++ skrl/utils/model_instantiators/torch/common.py
@@ -1,6 +1,29 @@
 from typing import Any, Mapping, Sequence, Tuple, Union
 
 import ast
+from gymnasium import spaces
+
+import torch
+import torch.nn as nn
+
+
+def one_hot_encoding(space: spaces.Space, x: torch.Tensor) -> torch.Tensor:
+    """One-hot encode a tensorized Discrete or MultiDiscrete space.
+
+    :param space: Gymnasium space
+    :param x: Tensorized sample/value of the given space
+
+    :return: One-hot encoded tensor
+    """
+    if isinstance(space, spaces.Discrete):
+        return nn.functional.one_hot(x[:, 0].long(), num_classes=space.n).float()
+    elif isinstance(space, spaces.MultiDiscrete):
+        return torch.cat(
+            [nn.functional.one_hot(x[:, i].long(), num_classes=space.nvec[i]) for i in range(space.nvec.shape[0])],
+            dim=1,
+        ).float()
+    else:
+        raise ValueError(f"Unsupported space ({space})")
 
 
 def _get_activation_function(activation: Union[str, None], as_module: bool = True) -> Union[str, None]:
@@ -62,6 +85,9 @@ def _parse_input(source: str) -> str:
     NodeTransformer().visit(tree)
     source = ast.unparse(tree)
     # enum substitutions
+    source = source.replace("OBSERVATION_SPACE", "self.observation_space")
+    source = source.replace("STATE_SPACE", "self.state_space")
+    source = source.replace("ACTION_SPACE", "self.action_space")
     source = source.replace("OBSERVATIONS", "observations")
     source = source.replace("STATES", "states")
     source = source.replace("ACTIONS", "taken_actions")