+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [2.0.0] - Unreleased
+## [1.4.2] - 2025-03-18
 ### Added
-- Allow ``None`` type spaces and samples/values in spaces utilities
+- Add Multi-Categorical model instantiator
+- Add `one_hot_encoding` function to model instantiators to one-hot encode `Discrete` and `MultiDiscrete` tensorized spaces
+- Allow `None` type spaces and samples/values in spaces utilities
+
+### Fixed
+- Cast model instantiator's `initial_log_std` parameter to `float` in PyTorch
+- Fix common property overwriting (e.g. `clip_actions`) in shared models composed of different mixin types
 
 ## [1.4.1] - 2025-01-27
 ### Fixed
@@ -97,7 +103,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.1.0] - 2024-02-12
 ### Added
-- MultiCategorical mixin to operate MultiDiscrete action spaces
+- `MultiCategoricalMixin` to operate `MultiDiscrete` action spaces
 
 ### Changed (breaking changes)
 - Rename the `ManualTrainer` to `StepTrainer`